requirements.txt
django
djangorestframework
sastrawi
transformers
gunicorn
